Fixed faulty values for total energy metrics. Fixes #8
